% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hello_world}
\alias{hello_world}
\title{Say hello!}
\usage{
hello_world()
}
\value{
Logical
}
\description{
Run \code{echo "Hello World!"} in Ubuntu 18.04.
}
\examples{
library(outsider)
hello_world <- module_import('hello_world',
                             repo = 'dombennett/om..hello.world')
hello_world()
}
